move scale note utils to utils and implement ScaleInfo
