+++ src/lib/ScaleInfo.svelte
@@ -2,8 +2,20 @@
   import { onMount } from "svelte";
   import { scalePoint, range } from "d3";
   import { Note } from "tonal";
+  import FretNote from "./FretNote.svelte";
 
   let { scale } = $props();
+
+  let notes = $derived(
+    scale.notes.map((note, index) => {
+      const noteObj = Note.get(note);
+      return {
+        ...noteObj,
+        label: noteObj.name.replace("b", "♭").replace("#", "♯"),
+        interval: scale.intervals[index],
+      };
+    }),
+  );
 
   let container = $state(null);
   let width = $state(0);
@@ -16,8 +28,6 @@
       .range([margin.left, width - margin.right]),
   );
 
-  const notes = $derived(scale.notes.map((n) => Note.get(n)));
-
   onMount(() => {
     width = container.clientWidth;
     height = 45;
@@ -28,21 +38,13 @@
   <svg class="h-12" viewBox="0 0 {width} {height}">
     {#each notes as note, i}
       <g transform="translate({dotX(i)}, {20})">
-        <circle
-          r="12"
-          class={note.interval === "1P" ? "fill-red-500" : "fill-green-500"} />
-        <text
-          text-anchor="middle"
-          dy="4"
-          font-size="10"
-          class="text-white"
-          fill="currentColor">{note.name}</text>
+        <FretNote {note} />
         <text
           text-anchor="middle"
           dy="25"
           font-size="10"
           class="text-black"
-          fill="currentColor">{scale.intervals[i]}</text>
+          fill="currentColor">{note.interval}</text>
       </g>
     {/each}
   </svg>
