+++ src/lib/utils.js
@@ -1,3 +1,5 @@
+import { Note } from "tonal";
+
 export const delay = (duration, value) => {
   return new Promise(function (resolve) {
     setTimeout(function () {
@@ -5,3 +7,20 @@ export const delay = (duration, value) => {
     }, duration);
   });
 };
+
+// Create a map of note names to their scale intervals
+export function getScaleNoteIntervalMap(scale) {
+  return scale.notes.reduce((map, note, index) => {
+    map[note] = scale.intervals[index];
+    return map;
+  }, {});
+}
+
+// Create a map of note chromas to scale note names for label normalization
+export function getScaleNoteLabelMap(scale) {
+  return scale.notes.reduce((map, note) => {
+    const noteObj = Note.get(note);
+    map[noteObj.chroma] = note;
+    return map;
+  }, {});
+}
